requirements.txt
azure_storage==0.37.0
msrest==0.6.21
pandas==1.4.4
python-dotenv==0.21.0
